pp_fastapi/requirements.txt
aiohttp==3.7.4.post0
aiosqlite==0.16.1
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
ciso8601==2.1.3
fastapi==0.67.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.1
idna==3.2
iso8601==0.1.16
multidict==5.1.0
priority==2.0.0
pydantic==1.8.2
pypika-tortoise==0.1.1
python-rapidjson==1.4
pytz==2021.1
requests==2.26.0
starlette==0.14.2
toml==0.10.2
tortoise-orm==0.17.5
typing-extensions==3.10.0.0
urllib3==1.26.6
uvloop==0.14.0
wsproto==1.0.0
yarl==1.6.3
